BUG: Avoid 32-bit index/offset overflow in CUDA filters for large images
Use 64-bit types for pointer offsets and element counts in the CUDA
filters so that reconstructions on large projection stacks or volumes
do not overflowing signed 32-bit arithmetic (which caused illegal
memory access / segmentation faults).

+++ src/rtkCudaAverageOutOfROIImageFilter.cu
@@ -47,7 +47,7 @@ average_along_dim_4(float * in, float * out, float * roi, unsigned int strideInF
     return;
 
   // Compute the index of the initial voxel
-  long int id = (k * c_Size.y + j) * c_Size.x + i;
+  long int id = (static_cast<long int>(k) * c_Size.y + j) * c_Size.x + i;
   long int strided_id = id; // strided_id will run along the 4th dimension
 
   // Compute the average along last dimension
