+++ src/rtkCudaFirstOrderKernels.cu
@@ -28,10 +28,10 @@ divergence_kernel(float * grad_x, float * grad_y, float * grad_z, float * out, i
   if (i >= c_Size.x || j >= c_Size.y || k >= c_Size.z)
     return;
 
-  long int id = (k * c_Size.y + j) * c_Size.x + i;
-  long int id_x = (k * c_Size.y + j) * c_Size.x + i - 1;
-  long int id_y = (k * c_Size.y + j - 1) * c_Size.x + i;
-  long int id_z = ((k - 1) * c_Size.y + j) * c_Size.x + i;
+  long int id = (static_cast<long int>(k) * c_Size.y + j) * c_Size.x + i;
+  long int id_x = (static_cast<long int>(k) * c_Size.y + j) * c_Size.x + i - 1;
+  long int id_y = (static_cast<long int>(k) * c_Size.y + j - 1) * c_Size.x + i;
+  long int id_z = ((static_cast<long int>(k) - 1) * c_Size.y + j) * c_Size.x + i;
 
   float3 A;
   float3 B;
@@ -76,10 +76,10 @@ gradient_kernel(float * in, float * grad_x, float * grad_y, float * grad_z, int3
   if (i >= c_Size.x || j >= c_Size.y || k >= c_Size.z)
     return;
 
-  long int id = (k * c_Size.y + j) * c_Size.x + i;
-  long int id_x = (k * c_Size.y + j) * c_Size.x + i + 1;
-  long int id_y = (k * c_Size.y + j + 1) * c_Size.x + i;
-  long int id_z = ((k + 1) * c_Size.y + j) * c_Size.x + i;
+  long int id = (static_cast<long int>(k) * c_Size.y + j) * c_Size.x + i;
+  long int id_x = (static_cast<long int>(k) * c_Size.y + j) * c_Size.x + i + 1;
+  long int id_y = (static_cast<long int>(k) * c_Size.y + j + 1) * c_Size.x + i;
+  long int id_z = ((static_cast<long int>(k) + 1) * c_Size.y + j) * c_Size.x + i;
 
   if (i == (c_Size.x - 1))
     grad_x[id] = 0;
