+++ src/rtkCudaLastDimensionTVDenoisingImageFilter.cu
@@ -45,8 +45,9 @@ denoise_oneD_TV_kernel(float * in, float * out, float beta, float gamma, int nit
   float *                 gradient = &shared[3 * c_Size.w];
 
   // Each thread reads one element into the shared buffer
-  long int gindex = ((threadIdx.x * c_Size.z + blockIdx.z) * c_Size.y + blockIdx.y) * c_Size.x + blockIdx.x;
-  int      lindex = threadIdx.x;
+  long int gindex =
+    ((static_cast<long int>(threadIdx.x) * c_Size.z + blockIdx.z) * c_Size.y + blockIdx.y) * c_Size.x + blockIdx.x;
+  int lindex = threadIdx.x;
   input[lindex] = in[gindex];
   __syncthreads();
